+++ examples/pytorch/graphsage/test_ray.py
@@ -5,6 +5,7 @@ import pytest
 import sys
 import os
 import platform
+import tempfile
 from typing import Dict
 
 import numpy as np
@@ -87,7 +88,7 @@ def train_func(config: Dict):
 
     dataset = ray.data.range(2708, parallelism=1)
     pipe = dataset.window(blocks_per_window=2).repeat(config["epochs"])
-    g = Client("/tmp/cora", [0], delayed_start=True)
+    g = Client(config.data_dir, [0], delayed_start=True)
 
     def transform_batch(batch: list) -> dict:
         return NeuralNetwork.query(g, batch)
@@ -112,14 +113,23 @@ def train_func(config: Dict):
 
 
 def test_graphsage_ppi_hvd_trainer():
+    working_dir = tempfile.TemporaryDirectory()
+    Cora(working_dir.name)
+
     ray.init()
     trainer = TorchTrainer(
         train_func,
-        train_loop_config={"lr": 1e-3, "batch_size": 64, "epochs": 4},
-        scaling_config=ScalingConfig(num_workers=2, use_gpu=False),
+        train_loop_config={
+            "data_dir": working_dir.name,
+            "lr": 1e-3,
+            "batch_size": 64,
+            "epochs": 4,
+        },
+        scaling_config=ScalingConfig(num_workers=1, use_gpu=False),
     )
     result = trainer.fit()
     print(f"Results: {result.metrics}")
+    working_dir.cleanup()
 
 
 if __name__ == "__main__":