+++ src/cc/lib/py_graph.h
@@ -76,11 +76,12 @@ extern "C"
     typedef void (*GetSparseFeaturesCallback)(const int64_t **, size_t *, const uint8_t **, size_t *, int64_t *);
     typedef void (*GetStringFeaturesCallback)(size_t, const uint8_t *);
 
-    DEEPGNN_DLL extern int32_t CreateLocalGraph(PyGraph *graph, size_t count, uint32_t *partitions,
-                                                const char *filename, PyPartitionStorageType storage_type,
-                                                const char *config_path);
+    DEEPGNN_DLL extern int32_t CreateLocalGraph(PyGraph *graph, const char *meta_location, size_t count,
+                                                uint32_t *partition_indices, const char **partition_locations,
+                                                PyPartitionStorageType storage_type, const char *config_path);
 
-    DEEPGNN_DLL extern int32_t StartServer(PyServer *graph, size_t count, uint32_t *partitions, const char *filename,
+    DEEPGNN_DLL extern int32_t StartServer(PyServer *graph, const char *meta_location, size_t count,
+                                           uint32_t *partition_indices, const char **partition_locations,
                                            const char *host_name, const char *ssl_key, const char *ssl_cert,
                                            const char *ssl_root, const PyPartitionStorageType storage_type,
                                            const char *config_path);