+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Breaking. Rename get_feature_type -> get_python_type.
+
 ## [0.1.56] - 2022-11-02
 
 ### Added